==> ALSA/6_Physical_measures/1_weight.Rmd <==
---
title: "Weight"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable `weight`

The description of harmonised variable is the following:
  
* Name: `weight`
* Label: `weight of the participant in kg`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
*Comments: `Weight should be collected in kilograms. It should be determined wheather the  variable was collected through a self reported measure or an athropometric test.`



### ALSA - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WEIGHT`
* Label: `Weight in kilograms`
* Type of variable: `countinous`
* Missings
    + `999 = Missing`
    
```{r assign1}
opal.assign.table.tibble(o, 'ALSA_w1_weig','ALSA.ALSA-Wave1',
                         variables=list('WEIGHT'), missings = TRUE)
```

```{r local1}
ALSA_w1_weig <- opal.execute(o,'ALSA_w1_weig')
Continuous_summary(var = ALSA_w1_weig$WEIGHT, missing_values = c(999))
ggplot(ALSA_w1_weig, aes(WEIGHT)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `keep missing into 999`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo1}
ALSA_ds_w1 <- tibble(id=ALSA_w1_weig$id)
ALSA_ds_w1$weight <- car::recode(ALSA_w1_weig$WEIGHT, "NA=999")
ALSA_ds_w1$weight <- labelled(ALSA_ds_w1$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1}
Continuous_summary(var = ALSA_ds_w1$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,180)
```

#### Validation
```{r crosstabulation1}
```




### ALSA - Wave 2 

There is no variable.ol




### ALSA - Wave 3 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WEIGHTW3`
* Label: `WEIGHT IN KILOGRAMS`
* Type of variable: `countinous`
* Missings
    + `999 = Missing`
    
```{r assign3}
opal.assign.table.tibble(o, 'ALSA_w3_weig','ALSA.ALSA-Wave3',
                         variables=list('WEIGHTW3'), missings = TRUE)

```

```{r local3}
ALSA_w3_weig <- opal.execute(o,'ALSA_w3_weig')
Continuous_summary(var = ALSA_w3_weig$WEIGHTW3, missing_values = c(999))
ggplot(ALSA_w3_weig, aes(WEIGHTW3)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(30,180)
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `keep missing into 999`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo3}
ALSA_ds_w3 <- tibble(id=ALSA_w3_weig$id)
ALSA_ds_w3$weight <- car::recode(ALSA_w3_weig$WEIGHTW3, "NA=999")
ALSA_ds_w3$weight <- labelled(ALSA_ds_w3$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript3}
Continuous_summary(var = ALSA_ds_w3$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w3, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,180)
```

#### Validation
```{r crosstabulation3}
```






### ALSA - Wave 4 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WEIGHKW4`
* Label: `WEIGHT IN KILOGRAMS`
* Type of variable: `countinous`
* Missings
    + `-1 = Dont know`
    + `1 = Same as last time`
    
```{r assign4}
opal.assign.table.tibble(o, 'ALSA_w4_weig','ALSA.ALSA-Wave4',
                         variables=list('WEIGHKW4'), missings = TRUE)
```

```{r local4}
ALSA_w4_weig <- opal.execute(o,'ALSA_w4_weig')
Continuous_summary(var = ALSA_w4_weig$WEIGHKW4, missing_values = c(-1,1))
ggplot(ALSA_w4_weig, aes(WEIGHKW4)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

There are values 1 that are going to be considered as mistakes instead of -1. Therefore it will be car::recoded as -1 value.

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `car::recode -1 "Dont know" into 998`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo4}
ALSA_ds_w4 <- tibble(id=ALSA_w4_weig$id)
ALSA_ds_w4$weight <- car::recode(ALSA_w4_weig$WEIGHKW4, "NA=999; -1=998; 1=998")
ALSA_ds_w4$weight <- labelled(ALSA_ds_w4$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4}
Continuous_summary(var = ALSA_ds_w4$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w4, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,183)
```

#### Validation
```{r crosstabulation4}
```




### ALSA - Wave 5 

There is no variable.






### ALSA - Wave 6 

There is no variable.






### ALSA - Wave 7 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `weightw7`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `999 = not attempted/done`
    
```{r assign7}
opal.assign.table.tibble(o, 'ALSA_w7_weig','ALSA.ALSA-Wave7',
                         variables=list('weightw7'), missings = TRUE)
```

```{r local7}
ALSA_w7_weig <- opal.execute(o,'ALSA_w7_weig')
Continuous_summary(var = ALSA_w7_weig$weightw7, missing_values = c(999))
ggplot(ALSA_w7_weig, aes(weightw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `car::recode 999 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo7}
ALSA_ds_w7 <- tibble(id=ALSA_w7_weig$id)
ALSA_ds_w7$weight <- car::recode(ALSA_w7_weig$weightw7, "999=996; NA=999")
ALSA_ds_w7$weight <- labelled(ALSA_ds_w7$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript7}
Continuous_summary(var = ALSA_ds_w7$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w7, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,115)
```

#### Validation
```{r crosstabulation7}
```




### ALSA - Wave 8 

There is no variable.






### ALSA - Wave 9 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WEIGHTW9`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
```{r assign9}
opal.assign.table.tibble(o, 'ALSA_w9_weig','ALSA.ALSA-Wave9',
                         variables=list('WEIGHTW9'), missings = TRUE)

```

```{r local9}
ALSA_w9_weig <- opal.execute(o,'ALSA_w9_weig')
Continuous_summary(var = ALSA_w9_weig$WEIGHTW9, missing_values = c(99.9))
ggplot(ALSA_w9_weig, aes(WEIGHTW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `car::recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo9}
ALSA_ds_w9 <- tibble(id=ALSA_w9_weig$id)
ALSA_ds_w9$weight <- car::recode(ALSA_w9_weig$WEIGHTW9, "99.9=996; NA=999")
ALSA_ds_w9$weight <- labelled(ALSA_ds_w9$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript9}
Continuous_summary(var = ALSA_ds_w9$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w9, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,180)
```

#### Validation
```{r crosstabulation9}
```





### ALSA - Wave 10 

There is no variable.





### ALSA - Wave 11 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WEIGHTW11`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
```{r assign11}
opal.assign.table.tibble(o, 'ALSA_w11_weig','ALSA.ALSA-Wave11',
                         variables=list('WEIGHTW11'), missings = TRUE)
```

```{r local11}
ALSA_w11_weig <- opal.execute(o,'ALSA_w11_weig')
Continuous_summary(var = ALSA_w11_weig$WEIGHTW11, missing_values = c(99.9))
ggplot(ALSA_w11_weig, aes(WEIGHTW11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `car::recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo11}
ALSA_ds_w11 <- tibble(id=ALSA_w11_weig$id)
ALSA_ds_w11$weight <- car::recode(ALSA_w11_weig$WEIGHTW11, "99.9=996; NA=999")
ALSA_ds_w11$weight <- labelled(ALSA_ds_w11$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript11}
Continuous_summary(var = ALSA_ds_w11$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w11, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(40,105)
```

#### Validation
```{r crosstabulation11}
```





### ALSA - Wave 12 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WEIGHTW12`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `999 = not attempted/done`
    
```{r assign12}
opal.assign.table.tibble(o, 'ALSA_w12_weig','ALSA.ALSA-Wave12',
                         variables=list('WEIGHTW12'), missings = TRUE)
```

```{r local12}
ALSA_w12_weig <- opal.execute(o,'ALSA_w12_weig')
Continuous_summary(var = ALSA_w12_weig$WEIGHTW12, missing_values = c(999))
ggplot(ALSA_w12_weig, aes(WEIGHTW12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(35,100)
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `car::recode 999 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo12}
ALSA_ds_w12 <- tibble(id=ALSA_w12_weig$id)
ALSA_ds_w12$weight <- car::recode(ALSA_w12_weig$WEIGHTW12, "999=996; NA=999")
ALSA_ds_w12$weight <- labelled(ALSA_ds_w12$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript12}
Continuous_summary(var = ALSA_ds_w12$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w12, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,180)
```

#### Validation
```{r crosstabulation12}
```




### ALSA - Wave 13 

There is no variable.




# Quality control
Anthropometric measure.



```{r save, echo=FALSE}
weight <- ALSA_ds_w1
save(weight, file = "../RData/w1/weight.RData")
rm(weight)

weight <- ALSA_ds_w3
save(weight, file = "../RData/w3/weight.RData")
rm(weight)

weight <- ALSA_ds_w4
save(weight, file = "../RData/w4/weight.RData")
rm(weight)

weight <- ALSA_ds_w7
save(weight, file = "../RData/w7/weight.RData")
rm(weight)

weight <- ALSA_ds_w9
save(weight, file = "../RData/w9/weight.RData")
rm(weight)

weight <- ALSA_ds_w11
save(weight, file = "../RData/w11/weight.RData")
rm(weight)

weight <- ALSA_ds_w12
save(weight, file = "../RData/w12/weight.RData")
rm(weight)

```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
