==> ALSA/6_Physical_measures/2_height.Rmd <==
---
title: "Height"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable `Height`

The description of harmonised variable is the following:
  
* Name: `height`
* Label: `height of the participant`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

* Comments: `Antropomeric measure for height should be measured in centimeters.`

# Data process

### ALSA - Wave 1 

There is no variable




### ALSA - Wave 2 

There is no variable.




### ALSA - Wave 3 

There is no variable.




### ALSA - Wave 4 

There is no variable.






### ALSA - Wave 5 

There is no variable.






### ALSA - Wave 6 

There is no variable.






### ALSA - Wave 7 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `heightw7`
* Label: `Height to nearest 0.1cm`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
```{r assign7}
opal.assign.table.tibble(o, 'ALSA_w7_heig','ALSA.ALSA-Wave7',
                         variables=list('heightw7'), missings = TRUE)
```

```{r local7}
ALSA_w7_heig <- opal.execute(o,'ALSA_w7_heig')
Continuous_summary(var = ALSA_w7_heig$heightw7, missing_values = c(99.9))
ggplot(ALSA_w7_heig, aes(heightw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") + xlim(100,200)
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `car::recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo7}
ALSA_ds_w7 <- tibble(id=ALSA_w7_heig$id)
ALSA_ds_w7$height <- car::recode(ALSA_w7_heig$heightw7, "NA=999; 99.9=996")
ALSA_ds_w7$height <- labelled(ALSA_ds_w7$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript7}
Continuous_summary(var = ALSA_ds_w7$height, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w7, aes(height)) + geom_histogram(stat="count", fill="steelblue") + xlab("height") + ylab("Frequency") + xlim(100,200) 
```

#### Validation
```{r crosstabulation7}
```








### ALSA - Wave 8 

There is no variable.






### ALSA - Wave 9 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `HEIGHTW9`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
```{r assign9}
opal.assign.table.tibble(o, 'ALSA_w9_heig','ALSA.ALSA-Wave9',
                         variables=list('HEIGHTW9'), missings = TRUE)
```

```{r local9}
ALSA_w9_heig <- opal.execute(o,'ALSA_w9_heig')
Continuous_summary(var = ALSA_w9_heig$HEIGHTW9, missing_values = c(99.9, 99.0))
ggplot(ALSA_w9_heig, aes(HEIGHTW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") + xlim(100,200)
```

There are 2 values 99 that are going to be considered as mistakes instead of 99.9. Therefore they will be car::recoded as 996. 

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be car::recoded as follows:

* `car::recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo9}
ALSA_ds_w9 <- tibble(id=ALSA_w9_heig$id)
ALSA_ds_w9$height <- car::recode(ALSA_w9_heig$HEIGHTW9, "NA=999; c(99,99.9)=996")
ALSA_ds_w9$height <- labelled(ALSA_ds_w9$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript9}
Continuous_summary(var = ALSA_ds_w9$height, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w9, aes(height)) + geom_histogram(stat="count", fill="steelblue") + xlab("height") + ylab("Frequency") + xlim(100,200) 
```

#### Validation
There is an unusual height value that is upper than 210 cm:
```{r crosstabulation9}
ALSA_ds_w9[which(ALSA_ds_w9$height>210 & ALSA_ds_w9$height<900),]
```






### ALSA - Wave 10 

There is no variable.





### ALSA - Wave 11 

There is no variable.





### ALSA - Wave 12 

There is no variable.





### ALSA - Wave 13 

There is no variable.


# Quality estimation
No comments.


```{r save, echo=FALSE}
height <- ALSA_ds_w7
save(height, file = "../RData/w7/height.RData")
rm(height)

height <- ALSA_ds_w9
save(height, file = "../RData/w9/height.RData")
rm(height)
```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
