==> ALSA/6_Physical_measures/3_bmi.Rmd <==
---
title: "BMI"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable `bmi`

The description of harmonised variable is the following:
  
* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
*Comments: `the BMI variable should be created by dividing weight and heigh variables with the following equation= weight (kg)/height(m)^2.`

# Data process

### ALSA - Wave 1 

There is no variable.




### ALSA - Wave 2 

There is no variable.




### ALSA - Wave 3 

There is no variable.



### ALSA - Wave 4 

There is no variable.






### ALSA - Wave 5 

There is no variable.






### ALSA - Wave 6 

There is no variable.






### ALSA - Wave 7 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `weightw7`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `999 = not attempted/done`

* Name: `heightw7`
* Label: `Height to nearest 0.1cm`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

    
```{r assign7}
opal.assign.table.tibble(o, 'ALSA_w7_bmi','ALSA.ALSA-Wave7',
                         variables=list('weightw7', 'heightw7'), missings = TRUE)
```

```{r local7}
ALSA_w7_bmi <- opal.execute(o,'ALSA_w7_bmi')

Continuous_summary(var = ALSA_w7_bmi$weightw7, missing_values = c(999))
ggplot(ALSA_w7_bmi, aes(weightw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(20,180)

Continuous_summary(var = ALSA_w7_bmi$heightw7, missing_values = c(99.9))
ggplot(ALSA_w7_bmi, aes(heightw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") + xlim(50,100)
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be car::recoded as follows:

* `car::recode 99.9 and 999 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo7}
compute_bmi <- function(Heigth, Weigth, miss, heigth_m = TRUE){
  Heigth <- car::recode(Heigth, "NA=999")
  Weigth <- car::recode(Weigth, "NA=999")
  
  if(heigth_m){
    bmi <- Weigth/(Heigth^2)
  
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
  } else {
    bmi <- Weigth/((Heigth/100)^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
  }
  bmi
}
ALSA_w7_bmi$bmi <- compute_bmi(Heigth = ALSA_w7_bmi$heightw7, 
                        Weigth = ALSA_w7_bmi$weightw7, miss = c(99.9), 
                        heigth_m = FALSE)


ALSA_ds_w7 <- tibble(id=ALSA_w7_bmi$id)
ALSA_ds_w7$bmi <- car::recode(ALSA_w7_bmi$bmi, "99.9=996; NA=999")
ALSA_ds_w7$bmi <- labelled(ALSA_ds_w7$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript7}
Continuous_summary(var = ALSA_ds_w7$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
 ggplot(ALSA_ds_w7, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
```{r crosstabulation7}
```




### ALSA - Wave 8 

There is no variable.






### ALSA - Wave 9 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WEIGHTW9`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

* Name: `HEIGHTW9`
* Label: `Weight to nearest kg`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

    
```{r assign9}
opal.assign.table.tibble(o, 'ALSA_w9_bmi','ALSA.ALSA-Wave9',
                         variables=list('WEIGHTW9', 'HEIGHTW9'), missings = TRUE)

```

```{r local9}
ALSA_w9_bmi <- opal.execute(o,'ALSA_w9_bmi')

Continuous_summary(var = ALSA_w9_bmi$WEIGHTW9, missing_values = c(99.9))
ggplot(ALSA_w9_bmi, aes(WEIGHTW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency") + xlim(20,180)

Continuous_summary(var = ALSA_w9_bmi$HEIGHTW9, missing_values = c(99.9))
ggplot(ALSA_w9_bmi, aes(HEIGHTW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") + xlim(50,80)
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be car::recoded as follows:

* `car::recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo9}
compute_bmi <- function(Heigth, Weigth, miss, heigth_m = TRUE){
  Heigth <- car::recode(Heigth, "NA=999")
  Weigth <- car::recode(Weigth, "NA=999")
  
  if(heigth_m){
    bmi <- Weigth/(Heigth^2)
  
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
  } else {
    bmi <- Weigth/((Heigth/100)^2)
    
    bmi[Heigth==999] <- 999
    bmi[Heigth==miss[1]] <- Heigth[Heigth==miss[1]]
    
    bmi[Weigth==999] <- 999
    bmi[Weigth==miss[1]] <- Weigth[Weigth==miss[1]]
  }
  bmi
}
ALSA_w9_bmi$bmi <- compute_bmi(Heigth = ALSA_w9_bmi$HEIGHTW9, 
                        Weigth = ALSA_w9_bmi$WEIGHTW9, miss = c(99.9), 
                        heigth_m = FALSE)


ALSA_ds_w9 <- tibble(id=ALSA_w9_bmi$id)
ALSA_ds_w9$bmi <- car::recode(ALSA_w9_bmi$bmi, "99.9=996; NA=999")
ALSA_ds_w9$bmi <- labelled(ALSA_ds_w9$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript9}
Continuous_summary(var = ALSA_ds_w9$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w9, aes(round(bmi,3))) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
There is an unusual bmi values, that is lower than 12
```{r crosstabulation9}
ALSA_w9_bmi[which(ALSA_w9_bmi$bmi<12),]
```






### ALSA - Wave 10 

There is no variable.





### ALSA - Wave 11 

There is no variable.



### ALSA - Wave 12 

There is no variable. 



### ALSA - Wave 13 

There is no variable.



# Quality estimation
No comments.


```{r save, echo=FALSE}
bmi <- ALSA_ds_w7
save(bmi, file = "../RData/w7/bmi.RData")
rm(bmi)

bmi <- ALSA_ds_w9
save(bmi, file = "../RData/w9/bmi.RData")
rm(bmi)
```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
