==> ALSA/6_Physical_measures/4_obesity.Rmd <==
---
title: "Obesity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable 

The description of harmonised variable is the following:

* Name: `obesity`
* Label: `Precense of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2.`
* Type of variable: `categorical`
* Categories: 
    + `0 = not obese`
    + `1 = obese`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing` 
*Comments: `the BMI variable should be created by dividing weight and hiegh variables by means of the following equation= weight (kg)/height(m)^2. Then, the Obesity variable should be generated according to the WHO cutoff point for BMI classification.`


```{r lab, echo=FALSE}
ds_label <- "Precense of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2"
```

### Data process

#### Australia

##### Wave 1

There is no variable.


##### Wave 2

There is no variable.


##### Wave 3

There is no variable.


##### Wave 4

There is no variable.


##### Wave 5

There is no variable.


##### Wave 6

There is no variable.


##### Wave 7

###### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
    
```{r assign_aksst1, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w7/bmi.RData")
ALSA_bmi <- bmi
```

```{r local_aksst1}
Continuous_summary(ALSA_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_bmi, aes(x=ALSA_bmi[[2]])) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,50)
```

###### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo1}
ALSA_ds_w7 <- tibble(id=ALSA_bmi$id)
ALSA_ds_w7$obesity <- car::recode(ALSA_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999")
ALSA_ds_w7$obesity <- labelled(ALSA_ds_w7[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

###### Statistical description of the new harmonised variable
```{r descript_aksst1}
Categorical_summary(var = ALSA_ds_w7[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w7, aes(x=factor(ALSA_ds_w7[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

###### Validation
```{r crosstab1}
BeforeH <- c(sum(na.omit(ALSA_bmi[[2]])<30),sum(na.omit(ALSA_bmi[[2]])>=30 & na.omit(ALSA_bmi[[2]])<=900), sum(na.omit(ALSA_bmi[[2]])==996), sum(na.omit(ALSA_bmi[[2]])==999))
AfterH <- table(ALSA_ds_w7[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","996", "999") 
kable(C)
```

```{r read1, include=FALSE}
rm(ALSA_bmi)
```

##### Wave 8

There is no variable.


##### Wave 9

###### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
    
```{r assign2, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w9/bmi.RData")
ALSA_bmi <- bmi
```

```{r local2}
Continuous_summary(ALSA_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_bmi, aes(ALSA_bmi[[2]])) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(8,70)
```

###### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `BMI>=1000 into 999`
* `NA into 999`

```{r harmo2}
ALSA_ds_w9 <- tibble(id=ALSA_bmi$id)
ALSA_ds_w9$obesity <- car::recode(ALSA_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999; 1000:hi=999")
ALSA_ds_w9$obesity <- labelled(ALSA_ds_w9[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

###### Statistical description of the new harmonised variable
```{r descript2}
Categorical_summary(var = ALSA_ds_w9[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w9, aes(x=factor(ALSA_ds_w9[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

###### Validation
```{r crosstab2}
BeforeH <- c(sum(na.omit(ALSA_bmi[[2]])<30),sum(na.omit(ALSA_bmi[[2]])>=30 & na.omit(ALSA_bmi[[2]])<=900), sum(na.omit(ALSA_bmi[[2]])==996), sum(na.omit(ALSA_bmi[[2]])==999))
AfterH <- table(ALSA_ds_w9[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","996", "999") 
kable(C)
```


```{r read2, include=FALSE}
rm(ALSA_bmi)
```

##### Wave 10

There is no variable.


##### Wave 11

There is no variable.

##### Wave 12

There is no variable.


##### Wave 13

There is no variable.


# Quality control
Once BMI has been calculated using weight (Kg) and Height (m^2),
Obesity was calculated by using the BMI cutoff>=30kg/m2 following the WHO 
classification. Finally, the resulting BMI variable was coded as follows: 
BMI>30kg/m2 =1 (obese); BMI<30kg/m2 =0 (not obese) to generate two categories.


```{r save, echo=FALSE}
obesity <- ALSA_ds_w7
save(obesity, file = "../RData/w7/obesity.RData")
rm(obesity)

obesity <- ALSA_ds_w9
save(obesity, file = "../RData/w9/obesity.RData")
rm(obesity)
```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
