==> ALSA/6_Physical_measures/5_waist.Rmd <==
---
title: "Waist"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable `waist`

The description of harmonised variable is the following:
  
* Name: `waist`
* Label: `waist circumference of the participant`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
     
* Comments: `Antropomeric measure for waist circumference should be measured in centimeters.`


### ALSA - Wave 1 

There is no variable. 


### ALSA - Wave 2 

There is no variable.ol


### ALSA - Wave 3 

There is no variable.


### ALSA - Wave 4 

There is no variable.


### ALSA - Wave 5 

There is no variable.


### ALSA - Wave 6 

There is no variable.



### ALSA - Wave 7 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `waistw7`
* Label: `Girth- waist, minimum (to nearest 0.1cm)`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
```{r assign7}
opal.assign.table.tibble(o, 'ALSA_w7_waist','ALSA.ALSA-Wave7',
                         variables=list('waistw7'), missings = TRUE)
```

```{r local7}
ALSA_w7_waist <- opal.execute(o,'ALSA_w7_waist')
Continuous_summary(var = ALSA_w7_waist$waistw7, missing_values = c(99.9))
ggplot(ALSA_w7_waist, aes(waistw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo7}
ALSA_ds_w7 <- tibble(id=ALSA_w7_waist$id)
ALSA_ds_w7$waist <- car::recode(ALSA_w7_waist$waistw7, "NA=999; 99.9=996")
ALSA_ds_w7$waist <- labelled(ALSA_ds_w7$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript7}
Continuous_summary(var = ALSA_ds_w7$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w7, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation7}
```



### ALSA - Wave 8 

There is no variable.






### ALSA - Wave 9 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WAIST1W9`
* Label: `Waist Girth 1`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
* Name: `WAIST2W9`
* Label: `Waist Girth 2`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

* Name: `WAIST3W9`
* Label: `Waist Girth 3`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

    
```{r assign9}
opal.assign.table.tibble(o, 'ALSA_w9_waist','ALSA.ALSA-Wave9',
                         variables=list('WAIST1W9', 'WAIST2W9', 'WAIST3W9'), missings = TRUE)
```

```{r local9}
ALSA_w9_waist <- opal.execute(o,'ALSA_w9_waist')

Continuous_summary(var = ALSA_w9_waist$WAIST1W9, missing_values = c(99.9, 999.9, 9999))
ggplot(ALSA_w9_waist, aes(WAIST1W9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_waist$WAIST2W9, missing_values = c(99.9, 999.9, 9999))
ggplot(ALSA_w9_waist, aes(WAIST2W9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_waist$WAIST3W9, missing_values = c(99.9, 999.9, 9999))
ggplot(ALSA_w9_waist, aes(WAIST3W9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

The variables have the values 999.9 and 9999 that are going to be considered as mistakes and will be recoded as 999 (missing). 

#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo9}
ALSA_w9_waist$mean_waist <- apply(X = ALSA_w9_waist[,c(2,3,4)], MARGIN = 1, FUN = mean)

ALSA_ds_w9 <- tibble(id=ALSA_w9_waist$id)
ALSA_ds_w9$waist <- car::recode(ALSA_w9_waist$mean_waist, "NA=999; 99.9=996; 999.9=999; 9999=999")
ALSA_ds_w9$waist <- labelled(ALSA_ds_w9$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript9}
Continuous_summary(var = ALSA_ds_w9$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w9, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation9}
```




### ALSA - Wave 10 

There is no variable.





### ALSA - Wave 11 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WAIST1W11`
* Label: `Waist Girth 1`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
* Name: `WAIST2W11`
* Label: `Waist Girth 2`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

* Name: `WAIST3W11`
* Label: `Waist Girth 3`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

    
```{r assign11}
opal.assign.table.tibble(o, 'ALSA_w11_waist','ALSA.ALSA-Wave11',
                         variables=list('WAIST1W11', 'WAIST2W11', 'WAIST3W11'), missings = TRUE)
```

```{r local11}
ALSA_w11_waist <- opal.execute(o,'ALSA_w11_waist')

Continuous_summary(var = ALSA_w11_waist$WAIST1W11, missing_values = c(99.9))
ggplot(ALSA_w11_waist, aes(WAIST1W11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_waist$WAIST2W11, missing_values = c(99.9))
ggplot(ALSA_w11_waist, aes(WAIST2W11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_waist$WAIST3W11, missing_values = c(99.9))
ggplot(ALSA_w11_waist, aes(WAIST3W11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `recode 99.9 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo11}
ALSA_w11_waist$mean_waist <- apply(X = ALSA_w11_waist[,c(2,3,4)], MARGIN = 1, FUN = mean)

ALSA_ds_w11 <- tibble(id=ALSA_w11_waist$id)
ALSA_ds_w11$waist <- car::recode(ALSA_w11_waist$mean_waist, "NA=999; 99.9=996")
ALSA_ds_w11$waist <- labelled(ALSA_ds_w11$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript11}
Continuous_summary(var = ALSA_ds_w11$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w11, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation11}
```




### ALSA - Wave 12 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `WAIST1W12`
* Label: `Waist Girth 1`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
* Name: `WAIST2W12`
* Label: `Waist Girth 2`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`

* Name: `WAIST3W12`
* Label: `Waist Girth 3`
* Type of variable: `countinous`
* Missings
    + `99.9 = not attempted/done`
    
```{r assign12}
opal.assign.table.tibble(o, 'ALSA_w12_waist','ALSA.ALSA-Wave12',
                         variables=list('WAIST1W12', 'WAIST2W12', 'WAIST3W12'), missings = TRUE)
```

```{r local12}
ALSA_w12_waist <- opal.execute(o,'ALSA_w12_waist')

Continuous_summary(var = ALSA_w12_waist$WAIST1W12, missing_values = c(99.9, 999.9))
ggplot(ALSA_w12_waist, aes(WAIST1W12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_waist$WAIST2W12, missing_values = c(99.9, 999.9))
ggplot(ALSA_w12_waist, aes(WAIST2W12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_waist$WAIST3W12, missing_values = c(99.9, 999.9))
ggplot(ALSA_w12_waist, aes(WAIST3W12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")

id999 <- ALSA_w12_waist$id[which(ALSA_w12_waist$WAIST3W12==999.9)]
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `recode 999 "not attempted/done" into 996`
* `NA as 999`

R code of the ALGORITHM:

```{r harmo12}
ALSA_w12_waist$mean_waist <- apply(X = ALSA_w12_waist[,c(2,3,4)], MARGIN = 1, FUN = mean)

ALSA_ds_w12 <- tibble(id=ALSA_w12_waist$id)
ALSA_ds_w12$waist <- car::recode(ALSA_w12_waist$mean_waist, "NA=999; 99.9=996; 999.9=999")
ALSA_ds_w12$waist[ALSA_ds_w12$id==id999[1]] <- ALSA_w12_waist$WAIST1W12[ALSA_w12_waist$id==id999[1]]
ALSA_ds_w12$waist <- labelled(ALSA_ds_w12$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript12}
Continuous_summary(var = ALSA_ds_w12$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w12, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation12}
```



### ALSA - Wave 13 

There is no variable.



#### Quality estimation
No coments.





```{r save, echo=FALSE}
waist <- ALSA_ds_w7
save(waist, file = "../RData/w7/waist.RData")
rm(waist)

waist <- ALSA_ds_w9
save(waist, file = "../RData/w9/waist.RData")
rm(waist)

waist <- ALSA_ds_w11
save(waist, file = "../RData/w11/waist.RData")
rm(waist)

waist <- ALSA_ds_w12
save(waist, file = "../RData/w12/waist.RData")
rm(waist)
```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
