==> ALSA/6_Physical_measures/10_sbp.Rmd <==
---
title: "Systolic blood pressure"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable `sbp`

The description of harmonised variable is the following:
  
* Name: `sbp`
* Label: `Systolic blood pressure of participant at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
*Comments: `Means of all sistolic blood pressure measures.`


### ALSA - Wave 1 

There is no variable.




### ALSA - Wave 2 

There is no variable.ol




### ALSA - Wave 3 

There is no variable. 




### ALSA - Wave 4 

There is no variable.





### ALSA - Wave 5 

There is no variable.






### ALSA - Wave 6 

There is no variable.






### ALSA - Wave 7 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `sbp15w7`
* Label: `Standing blood pressure- BP systolic at 45 sec`
* Type of variable: `countinous`

* Name: `sbp45w7`
* Label: `Standing blood pressure- BP systolic at 1 min 45 sec`
* Type of variable: `countinous`

* Name: `sbp3mw7`
* Label: `Sitting blood pressure-BP systolic at 3 min rest`
* Type of variable: `countinous`

* Name: `sbp4mw7`
* Label: `Sitting blood pressure-BP systolic at 4 min rest`
* Type of variable: `countinous`

* Name: `sbp5mw7`
* Label: `Sitting blood pressure- BP systolic at 5 min rest`
* Type of variable: `countinous`

```{r assign7}
opal.assign.table.tibble(o, 'ALSA_w7_sbp','ALSA.ALSA-Wave7',
                         variables=list('sbp15w7', 'sbp45w7', 'sbp3mw7', 'sbp4mw7', 'sbp5mw7'), missings = TRUE)
```

```{r local7}
ALSA_w7_sbp <- opal.execute(o,'ALSA_w7_sbp')

Continuous_summary(var = ALSA_w7_sbp$sbp15w7, missing_values = NA)
ggplot(ALSA_w7_sbp, aes(sbp15w7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_sbp$sbp45w7, missing_values = NA)
ggplot(ALSA_w7_sbp, aes(sbp45w7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_sbp$sbp3mw7, missing_values = NA)
ggplot(ALSA_w7_sbp, aes(sbp3mw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_sbp$sbp4mw7, missing_values = NA)
ggplot(ALSA_w7_sbp, aes(sbp4mw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_sbp$sbp5mw7, missing_values = NA)
ggplot(ALSA_w7_sbp, aes(sbp5mw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

ALSA_w7_sbp[which(ALSA_w7_sbp$sbp15w7==0),]
ind_0 <- which(ALSA_w7_sbp$sbp15w7==0)
```

There is a individual that has value 0 in the measurament 'sbp15w7' and 'sbp45w7'. The mean of this individual will be computed witout this values. 

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo7}
ALSA_w7_sbp$mean_sbp <- apply(X = ALSA_w7_sbp[,2:6], MARGIN = 1, FUN = mean)

ALSA_w7_sbp$mean_sbp[ind_0] <- apply(X = ALSA_w7_sbp[ind_0,2:4], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w7_sbp$sbp3mw7) | is.na(ALSA_w7_sbp$sbp4mw7) | 
                   is.na(ALSA_w7_sbp$sbp5mw7) | is.na(ALSA_w7_sbp$sbp15w7) | 
                   is.na(ALSA_w7_sbp$sbp45w7) ) 
all_Nas_ind <- which(is.na(ALSA_w7_sbp$sbp3mw7) & 
                       is.na(ALSA_w7_sbp$sbp4mw7) & 
                       is.na(ALSA_w7_sbp$sbp5mw7) & 
                       is.na(ALSA_w7_sbp$sbp15w7) & 
                       is.na(ALSA_w7_sbp$sbp45w7) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in sbp4mw7 and sbp5mw7:
new_Nas_ind_4_5 <- new_Nas_ind[2]
ALSA_w7_sbp$mean_sbp[new_Nas_ind_4_5] <- apply(X = ALSA_w7_sbp[new_Nas_ind_4_5,c(2,5,6)], MARGIN = 1, FUN = mean)

#individuals with NA in sbp45w7:
new_Nas_ind_5 <- new_Nas_ind[c(11, 13, 19)]
ALSA_w7_sbp$mean_sbp[new_Nas_ind_5] <- apply(X = ALSA_w7_sbp[new_Nas_ind_5,c(2:5)], MARGIN = 1, FUN = mean)

#individuals with NA in sbp15w7 and sbp45w7:
new_Nas_ind_15_45 <- new_Nas_ind[-c(2,11, 13, 19)]
ALSA_w7_sbp$mean_sbp[new_Nas_ind_15_45] <- apply(X = ALSA_w7_sbp[new_Nas_ind_15_45,c(2:4)], MARGIN = 1, FUN = mean)


ALSA_ds_w7 <- tibble(id=ALSA_w7_sbp$id)
ALSA_ds_w7$sbp <- car::recode(ALSA_w7_sbp$mean_sbp, "NA=999")
ALSA_ds_w7$sbp <- labelled(ALSA_ds_w7$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript7}
Continuous_summary(var = ALSA_ds_w7$sbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w7, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation7}
```



### ALSA - Wave 8 

There is no variable.






### ALSA - Wave 9 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `SBP15W9`
* Label: `BP- Systolic at 45 secs standing`
* Type of variable: `countinous`

* Name: `SBP45W9`
* Label: `BP- Systolic at 105 sec standing`
* Type of variable: `countinous`

* Name: `SBP3MW9`
* Label: `BP- Systolic at 3 mins`
* Type of variable: `countinous`

* Name: `SBP4MW9`
* Label: `BP- Systolic at 4 mins`
* Type of variable: `countinous`

* Name: `SBP5MW9`
* Label: `BP- Systolic at 5 mins`
* Type of variable: `countinous`
    
```{r assign9}
opal.assign.table.tibble(o, 'ALSA_w9_sbp','ALSA.ALSA-Wave9',
                         variables=list('SBP15W9', 'SBP45W9', 'SBP3MW9', 'SBP4MW9', 'SBP5MW9'), missings = TRUE)

```

```{r local9}
ALSA_w9_sbp <- opal.execute(o,'ALSA_w9_sbp')

Continuous_summary(var = ALSA_w9_sbp$SBP15W9, missing_values = NA)
ggplot(ALSA_w9_sbp, aes(SBP15W9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_sbp$SBP45W9, missing_values = NA)
ggplot(ALSA_w9_sbp, aes(SBP45W9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_sbp$SBP3MW9, missing_values = NA)
ggplot(ALSA_w9_sbp, aes(SBP3MW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_sbp$SBP4MW9, missing_values = NA)
ggplot(ALSA_w9_sbp, aes(SBP4MW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_sbp$SBP5MW9, missing_values = NA)
ggplot(ALSA_w9_sbp, aes(SBP5MW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo9}
ALSA_w9_sbp$mean_sbp <- apply(X = ALSA_w9_sbp[,2:6], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w9_sbp$SBP3MW9) | is.na(ALSA_w9_sbp$SBP4MW9) | 
                   is.na(ALSA_w9_sbp$SBP5MW9) | is.na(ALSA_w9_sbp$SBP15W9) | 
                   is.na(ALSA_w9_sbp$SBP45W9) ) 
all_Nas_ind <- which(is.na(ALSA_w9_sbp$SBP3MW9) & 
                       is.na(ALSA_w9_sbp$SBP4MW9) & 
                       is.na(ALSA_w9_sbp$SBP5MW9) & 
                       is.na(ALSA_w9_sbp$SBP15W9) & 
                       is.na(ALSA_w9_sbp$SBP45W9) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in all measuraments except in SBP4MW9:
new_Nas_ind_4 <- new_Nas_ind[1]
ALSA_w9_sbp$mean_sbp[new_Nas_ind_4] <- ALSA_w9_sbp$SBP4MW9[new_Nas_ind_4]

#individuals with NA in SBP45W9:
new_Nas_ind_45 <- new_Nas_ind[2]
ALSA_w9_sbp$mean_sbp[new_Nas_ind_45] <- apply(X = ALSA_w9_sbp[new_Nas_ind_45,c(2:5)], MARGIN = 1, FUN = mean)

#individuals with NA in SBP4MW9, SBP5MW9 and SBP45W9:
new_Nas_ind_4_5_45 <- new_Nas_ind[4]
ALSA_w9_sbp$mean_sbp[new_Nas_ind_4_5_45] <- apply(X = ALSA_w9_sbp[new_Nas_ind_4_5_45,c(2,5)], MARGIN = 1, FUN = mean)

#individuals with NA in SBP15W9 and SBP45W9:
new_Nas_ind_15_45 <- new_Nas_ind[c(3,5,6)]
ALSA_w9_sbp$mean_sbp[new_Nas_ind_4_5_45] <- apply(X = ALSA_w9_sbp[new_Nas_ind_4_5_45,c(2:4)], MARGIN = 1, FUN = mean)


ALSA_ds_w9 <- tibble(id=ALSA_w9_sbp$id)
ALSA_ds_w9$sbp <- car::recode(ALSA_w9_sbp$mean_sbp, "NA=999")
ALSA_ds_w9$sbp <- labelled(ALSA_ds_w9$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript9}
Continuous_summary(var = ALSA_ds_w9$sbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w9, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation9}
```




### ALSA - Wave 10 

There is no variable.





### ALSA - Wave 11 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `SBP15W11`
* Label: `BP- Systolic at 45 secs standing`
* Type of variable: `countinous`

* Name: `SBP45W11`
* Label: `BP- Systolic at 105 sec standing`
* Type of variable: `countinous`

* Name: `SBP3MW11`
* Label: `BP- Systolic at 3 mins`
* Type of variable: `countinous`

* Name: `SBP4MW11`
* Label: `BP- Systolic at 4 mins`
* Type of variable: `countinous`

* Name: `SBP5MW11`
* Label: `BP- Systolic at 5 mins`
* Type of variable: `countinous`
    
```{r assign11}
opal.assign.table.tibble(o, 'ALSA_w11_sbp','ALSA.ALSA-Wave11',
                         variables=list('SBP15W11', 'SBP45W11', 'SBP3MW11', 'SBP4MW11', 'SBP5MW11'), missings = TRUE)
```

```{r local11}
ALSA_w11_sbp <- opal.execute(o,'ALSA_w11_sbp')

Continuous_summary(var = ALSA_w11_sbp$SBP15W11, missing_values = NA)
ggplot(ALSA_w11_sbp, aes(SBP15W11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_sbp$SBP45W11, missing_values = NA)
ggplot(ALSA_w11_sbp, aes(SBP45W11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_sbp$SBP3MW11, missing_values = NA)
ggplot(ALSA_w11_sbp, aes(SBP3MW11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_sbp$SBP4MW11, missing_values = NA)
ggplot(ALSA_w11_sbp, aes(SBP4MW11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_sbp$SBP5MW11, missing_values = NA)
ggplot(ALSA_w11_sbp, aes(SBP5MW11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo11}
ALSA_w11_sbp$mean_sbp <- apply(X = ALSA_w11_sbp[,2:6], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w11_sbp$SBP3MW11) | is.na(ALSA_w11_sbp$SBP4MW11) | 
                   is.na(ALSA_w11_sbp$SBP5MW11) | is.na(ALSA_w11_sbp$SBP15W11) | 
                   is.na(ALSA_w11_sbp$SBP45W11) ) 
all_Nas_ind <- which(is.na(ALSA_w11_sbp$SBP3MW11) & 
                       is.na(ALSA_w11_sbp$SBP4MW11) & 
                       is.na(ALSA_w11_sbp$SBP5MW11) & 
                       is.na(ALSA_w11_sbp$SBP15W11) & 
                       is.na(ALSA_w11_sbp$SBP45W11) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in SBP4MW11 and SBP5MW11:
new_Nas_ind_4_5 <- new_Nas_ind[1]
ALSA_w11_sbp$mean_sbp[new_Nas_ind_4_5] <- apply(X = ALSA_w11_sbp[new_Nas_ind_4_5,c(2,5,6)], MARGIN = 1, FUN = mean)

#individuals with NA in SBP45W11
new_Nas_ind_45 <- new_Nas_ind[4:5]
ALSA_w11_sbp$mean_sbp[new_Nas_ind_45] <- apply(X = ALSA_w11_sbp[new_Nas_ind_45,c(2:5)], MARGIN = 1, FUN = mean)

#individuals with NA in SBP15W11 and SBP45W11:
new_Nas_ind_15_45<- new_Nas_ind[c(2:3,6)]
ALSA_w11_sbp$mean_sbp[new_Nas_ind_15_45] <- apply(X = ALSA_w11_sbp[new_Nas_ind_15_45,c(2:4)], MARGIN = 1, FUN = mean)


ALSA_ds_w11 <- tibble(id=ALSA_w11_sbp$id)
ALSA_ds_w11$sbp <- car::recode(ALSA_w11_sbp$mean_sbp, "NA=999")
ALSA_ds_w11$sbp <- labelled(ALSA_ds_w11$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript11}
Continuous_summary(var = ALSA_ds_w11$sbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w11, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation11}
```




### ALSA - Wave 12 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `SBP15W12`
* Label: `BP- Systolic at 45 secs standing`
* Type of variable: `countinous`

* Name: `SBP45W12`
* Label: `BP- Systolic at 105 sec standing`
* Type of variable: `countinous`

* Name: `SBP3MW12`
* Label: `BP- Systolic at 3 mins`
* Type of variable: `countinous`

* Name: `SBP4MW12`
* Label: `BP- Systolic at 4 mins`
* Type of variable: `countinous`

* Name: `SBP5MW12`
* Label: `BP- Systolic at 5 mins`
* Type of variable: `countinous`
    
```{r assign12}
opal.assign.table.tibble(o, 'ALSA_w12_sbp','ALSA.ALSA-Wave12',
                         variables=list('SBP15W12', 'SBP45W12', 'SBP3MW12', 'SBP4MW12', 'SBP5MW12'), missings = TRUE)
```

```{r local12}
ALSA_w12_sbp <- opal.execute(o,'ALSA_w12_sbp')

Continuous_summary(var = ALSA_w12_sbp$SBP15W12, missing_values = NA)
ggplot(ALSA_w12_sbp, aes(SBP15W12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_sbp$SBP45W12, missing_values = NA)
ggplot(ALSA_w12_sbp, aes(SBP45W12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_sbp$SBP3MW12, missing_values = NA)
ggplot(ALSA_w12_sbp, aes(SBP3MW12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_sbp$SBP4MW12, missing_values = NA)
ggplot(ALSA_w12_sbp, aes(SBP4MW12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_sbp$SBP5MW12, missing_values = NA)
ggplot(ALSA_w12_sbp, aes(SBP5MW12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo12}
ALSA_w12_sbp$mean_sbp <- apply(X = ALSA_w12_sbp[,2:6], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w12_sbp$SBP3MW12) | is.na(ALSA_w12_sbp$SBP4MW12) | 
                   is.na(ALSA_w12_sbp$SBP5MW12) | is.na(ALSA_w12_sbp$SBP15W12) | 
                   is.na(ALSA_w12_sbp$SBP45W12) ) 
all_Nas_ind <- which(is.na(ALSA_w12_sbp$SBP3MW12) & 
                       is.na(ALSA_w12_sbp$SBP4MW12) & 
                       is.na(ALSA_w12_sbp$SBP5MW12) & 
                       is.na(ALSA_w12_sbp$SBP15W12) & 
                       is.na(ALSA_w12_sbp$SBP45W12) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in SBP15W12 and SBP45W12:
new_Nas_ind_15_45 <- new_Nas_ind[c(1,4,5)]
ALSA_w12_sbp$mean_sbp[new_Nas_ind_15_45] <- apply(X = ALSA_w12_sbp[new_Nas_ind_15_45,c(2:4)], MARGIN = 1, FUN = mean)

#individuals with NA in SBP5MW12:
new_Nas_ind_5 <- new_Nas_ind[2]
ALSA_w12_sbp$mean_sbp[new_Nas_ind_5] <- apply(X = ALSA_w12_sbp[new_Nas_ind_5,c(2:3, 5:6)], MARGIN = 1, FUN = mean)

#individuals with NA in SBP45W12:
new_Nas_ind_45 <- new_Nas_ind[3]
ALSA_w12_sbp$mean_sbp[new_Nas_ind_45] <- apply(X = ALSA_w12_sbp[new_Nas_ind_45,c(2:5)], MARGIN = 1, FUN = mean)

#individuals with NA in all measuraments except in SBP3MW12:
new_Nas_ind_3 <- new_Nas_ind[6]
ALSA_w12_sbp$mean_sbp[new_Nas_ind_3] <- ALSA_w12_sbp$SBP3MW12[new_Nas_ind_3]

ALSA_ds_w12 <- tibble(id=ALSA_w12_sbp$id)
ALSA_ds_w12$sbp <- car::recode(ALSA_w12_sbp$mean_sbp, "NA=999")
ALSA_ds_w12$sbp <- labelled(ALSA_ds_w12$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript12}
Continuous_summary(var = ALSA_ds_w12$sbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w12, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation12}
```



### ALSA - Wave 13 

There is no variable.

# Quality estimation
Systolic bp was calculated as the all 5 variables mean, 
because these variables were used in all waves of the study.


```{r save, echo=FALSE}
sbp <- ALSA_ds_w7
save(sbp, file = "../RData/w7/sbp.RData")
rm(sbp)

sbp <- ALSA_ds_w9
save(sbp, file = "../RData/w9/sbp.RData")
rm(sbp)

sbp <- ALSA_ds_w11
save(sbp, file = "../RData/w11/sbp.RData")
rm(sbp)

sbp <- ALSA_ds_w12
save(sbp, file = "../RData/w12/sbp.RData")
rm(sbp)
```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
