==> ALSA/6_Physical_measures/11_dbp.Rmd <==
---
title: "Diastolic blood pressure"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 

# Description of DataSchema variable `dbp`

The description of harmonised variable is the following:
  
* Name: `dbp`
* Label: `Diastolic blood pressure of participantat at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
*Comments: `Means of all diastolic blood pressure measures.`

### ALSA - Wave 1 

There is no variable.




### ALSA - Wave 2 

There is no variable.ol




### ALSA - Wave 3 

There is no variable. 




### ALSA - Wave 4 

There is no variable.





### ALSA - Wave 5 

There is no variable.






### ALSA - Wave 6 

There is no variable.






### ALSA - Wave 7 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `dbp15w7`
* Label: `Standing blood pressure- BP diastolic at 45 sec`
* Type of variable: `countinous`

* Name: `dbp45w7`
* Label: `Standing blood pressure- BP diastolic at 1 min 45 sec`
* Type of variable: `countinous`

* Name: `dbp3mw7`
* Label: `Sitting blood pressure-BP diastolic at 3 min rest`
* Type of variable: `countinous`

* Name: `dbp4mw7`
* Label: `Sitting blood pressure-BP diastolic at 4 min rest`
* Type of variable: `countinous`

* Name: `dbp5mw7`
* Label: `Sitting blood pressure- BP diastolic at 5 min rest`
* Type of variable: `countinous`

```{r assign7}
opal.assign.table.tibble(o, 'ALSA_w7_dbp','ALSA.ALSA-Wave7',
                         variables=list('dbp15w7', 'dbp45w7', 'dbp3mw7', 'dbp4mw7', 'dbp5mw7'), missings = TRUE)
```

```{r local7}
ALSA_w7_dbp <- opal.execute(o,'ALSA_w7_dbp')

Continuous_summary(var = ALSA_w7_dbp$dbp15w7, missing_values = 0)
ggplot(ALSA_w7_dbp, aes(dbp15w7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_dbp$dbp45w7, missing_values = c(0,1))
ggplot(ALSA_w7_dbp, aes(dbp45w7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_dbp$dbp3mw7, missing_values = c(0))
ggplot(ALSA_w7_dbp, aes(dbp3mw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_dbp$dbp4mw7, missing_values = c(0))
ggplot(ALSA_w7_dbp, aes(dbp4mw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w7_dbp$dbp5mw7, missing_values = 0)
ggplot(ALSA_w7_dbp, aes(dbp5mw7)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

ALSA_w7_dbp[which(ALSA_w7_dbp$dbp45w7==0 | ALSA_w7_dbp$dbp3mw7==0 | ALSA_w7_dbp$dbp4mw7==0 | ALSA_w7_dbp$dbp5mw7==0 | ALSA_w7_dbp$dbp15w7==0| ALSA_w7_dbp$dbp45w7==1),]

ind_0 <- which(ALSA_w7_dbp$dbp45w7==0 | ALSA_w7_dbp$dbp3mw7==0 | ALSA_w7_dbp$dbp4mw7==0 | ALSA_w7_dbp$dbp5mw7==0 | ALSA_w7_dbp$dbp15w7==0| ALSA_w7_dbp$dbp45w7==1)
```

There is a individual that has value 0 in the measurament in some measurements. The mean of this individual will be computed witout this values. 

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo7}
ALSA_w7_dbp$mean_dbp <- apply(X = ALSA_w7_dbp[,2:6], MARGIN = 1, FUN = mean)

ALSA_w7_dbp$mean_dbp[ind_0[1]] <- apply(X = ALSA_w7_dbp[ind_0[1],2:3], MARGIN = 1, FUN = mean)
ALSA_w7_dbp$mean_dbp[ind_0[2]] <- apply(X = ALSA_w7_dbp[ind_0[2],5:6], MARGIN = 1, FUN = mean)

ALSA_w7_dbp$mean_dbp[ind_0[3]] <- apply(X = ALSA_w7_dbp[ind_0[3],2:5], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w7_dbp$dbp3mw7) | is.na(ALSA_w7_dbp$dbp4mw7) | 
                   is.na(ALSA_w7_dbp$dbp5mw7) | is.na(ALSA_w7_dbp$dbp15w7) | 
                   is.na(ALSA_w7_dbp$dbp45w7) ) 
all_Nas_ind <- which(is.na(ALSA_w7_dbp$dbp3mw7) & 
                       is.na(ALSA_w7_dbp$dbp4mw7) & 
                       is.na(ALSA_w7_dbp$dbp5mw7) & 
                       is.na(ALSA_w7_dbp$dbp15w7) & 
                       is.na(ALSA_w7_dbp$dbp45w7) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in dbp4mw7:
new_Nas_ind_4 <- new_Nas_ind[2]
ALSA_w7_dbp$mean_dbp[new_Nas_ind_4] <- apply(X = ALSA_w7_dbp[new_Nas_ind_4,c(2,4:6)], MARGIN = 1, FUN = mean)

#individuals with NA in dbp45w7:
new_Nas_ind_45 <- new_Nas_ind[c(11, 13, 19)]
ALSA_w7_dbp$mean_dbp[new_Nas_ind_45] <- apply(X = ALSA_w7_dbp[new_Nas_ind_45,c(2:5)], MARGIN = 1, FUN = mean)

#individuals with NA in dbp15w7 and dbp45w7:
new_Nas_ind_15_45 <- new_Nas_ind[c(1,3:10,12,14:18,20:22)]
ALSA_w7_dbp$mean_dbp[new_Nas_ind_15_45] <- apply(X = ALSA_w7_dbp[new_Nas_ind_15_45,c(2:4)], MARGIN = 1, FUN = mean)


ALSA_ds_w7 <- tibble(id=ALSA_w7_dbp$id)
ALSA_ds_w7$dbp <- car::recode(ALSA_w7_dbp$mean_dbp, "NA=999")
ALSA_ds_w7$dbp <- labelled(ALSA_ds_w7$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript7}
Continuous_summary(var = ALSA_ds_w7$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w7, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
There are a individual with a unusual value in the measurement dbp45w7, that is lower than 25.
```{r crosstabulation7}
ALSA_w7_dbp[ind_0[2],]
```





### ALSA - Wave 8 

There is no variable.






### ALSA - Wave 9 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `DBP15W9`
* Label: `BP- Diastolic at 45 secs standing`
* Type of variable: `countinous`

* Name: `DBP45W9`
* Label: `BP- Diastolicat 105 sec standing`
* Type of variable: `countinous`

* Name: `DBP3MW9`
* Label: `BP- Diastolic at 3 mins`
* Type of variable: `countinous`

* Name: `DBP4MW9`
* Label: `BP- Diastolic at 4 mins`
* Type of variable: `countinous`

* Name: `dbp5mw7`
* Label: `BP - Diastolic at 5 mins`
* Type of variable: `countinous`
    
```{r assign9}
opal.assign.table.tibble(o, 'ALSA_w9_dbp','ALSA.ALSA-Wave9',
                         variables=list('DBP15W9', 'DBP45W9', 'DBP3MW9', 'DBP4MW9', 'DBP5MW9'), missings = TRUE)

```

```{r local9}
ALSA_w9_dbp <- opal.execute(o,'ALSA_w9_dbp')

Continuous_summary(var = ALSA_w9_dbp$DBP15W9, missing_values = NA)
ggplot(ALSA_w9_dbp, aes(DBP15W9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_dbp$DBP45W9, missing_values = NA)
ggplot(ALSA_w9_dbp, aes(DBP45W9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_dbp$DBP3MW9, missing_values = NA)
ggplot(ALSA_w9_dbp, aes(DBP3MW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_dbp$DBP4MW9, missing_values = NA)
ggplot(ALSA_w9_dbp, aes(DBP4MW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w9_dbp$DBP5MW9, missing_values = NA)
ggplot(ALSA_w9_dbp, aes(DBP5MW9)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo9}
ALSA_w9_dbp$mean_dbp <- apply(X = ALSA_w9_dbp[,2:6], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w9_dbp$DBP3MW9) | is.na(ALSA_w9_dbp$DBP4MW9) | 
                   is.na(ALSA_w9_dbp$DBP5MW9) | is.na(ALSA_w9_dbp$DBP15W9) | 
                   is.na(ALSA_w9_dbp$DBP45W9) ) 
all_Nas_ind <- which(is.na(ALSA_w9_dbp$DBP3MW9) & 
                       is.na(ALSA_w9_dbp$DBP4MW9) & 
                       is.na(ALSA_w9_dbp$DBP5MW9) & 
                       is.na(ALSA_w9_dbp$DBP15W9) & 
                       is.na(ALSA_w9_dbp$DBP45W9) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individual with value only in DBP4MW9:
new_Nas_ind_4 <- new_Nas_ind[1]
ALSA_w9_dbp$mean_dbp[new_Nas_ind_4] <- ALSA_w9_dbp$DBP4MW9[new_Nas_ind_4]

#individuals with NA in DBP45W9:
new_Nas_ind_45 <- new_Nas_ind[2]
ALSA_w9_dbp$mean_dbp[new_Nas_ind_45] <- apply(X = ALSA_w9_dbp[new_Nas_ind_45,c(2:5)], MARGIN = 1, FUN = mean)

#individuals with NA in DBP4MW9, DBP5MW9 and DBP45W9:
new_Nas_ind_4_5_45 <- new_Nas_ind[4]
ALSA_w9_dbp$mean_dbp[new_Nas_ind_4_5_45] <- apply(X = ALSA_w9_dbp[new_Nas_ind_4_5_45,c(2,5)], MARGIN = 1, FUN = mean)


#individuals with NA in DBP15W9 and DBP45W9:
new_Nas_ind_15_45 <- new_Nas_ind[c(3,5,6)]
ALSA_w9_dbp$mean_dbp[new_Nas_ind_15_45] <- apply(X = ALSA_w9_dbp[new_Nas_ind_15_45,c(2:4)], MARGIN = 1, FUN = mean)


ALSA_ds_w9 <- tibble(id=ALSA_w9_dbp$id)
ALSA_ds_w9$dbp <- car::recode(ALSA_w9_dbp$mean_dbp, "NA=999")
ALSA_ds_w9$dbp <- labelled(ALSA_ds_w9$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript9}
Continuous_summary(var = ALSA_ds_w9$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w9, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation9}
```




### ALSA - Wave 10 

There is no variable.





### ALSA - Wave 11 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `DBP15W11`
* Label: `BP- Diastolic at 45 secs standing`
* Type of variable: `countinous`

* Name: `DBP45W11`
* Label: `BP- Diastolicat 105 sec standing`
* Type of variable: `countinous`

* Name: `DBP3MW11`
* Label: `BP- Diastolic at 3 mins`
* Type of variable: `countinous`

* Name: `DBP4MW11`
* Label: `BP- Diastolic at 4 mins`
* Type of variable: `countinous`

* Name: `DBP5MW11`
* Label: `BP - Diastolic at 5 mins`
* Type of variable: `countinous`
    
```{r assign11}
opal.assign.table.tibble(o, 'ALSA_w11_dbp','ALSA.ALSA-Wave11',
                         variables=list('DBP15W11', 'DBP45W11', 'DBP3MW11', 'DBP4MW11', 'DBP5MW11'), missings = TRUE)
```

```{r local11}
ALSA_w11_dbp <- opal.execute(o,'ALSA_w11_dbp')

Continuous_summary(var = ALSA_w11_dbp$DBP15W11, missing_values = NA)
ggplot(ALSA_w11_dbp, aes(DBP15W11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_dbp$DBP45W11, missing_values = NA)
ggplot(ALSA_w11_dbp, aes(DBP45W11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_dbp$DBP3MW11, missing_values = NA)
ggplot(ALSA_w11_dbp, aes(DBP3MW11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_dbp$DBP4MW11, missing_values = NA)
ggplot(ALSA_w11_dbp, aes(DBP4MW11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w11_dbp$DBP5MW11, missing_values = NA)
ggplot(ALSA_w11_dbp, aes(DBP5MW11)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo11}
ALSA_w11_dbp$mean_dbp <- apply(X = ALSA_w11_dbp[,2:6], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w11_dbp$DBP3MW11) | is.na(ALSA_w11_dbp$DBP4MW11) | 
                   is.na(ALSA_w11_dbp$DBP5MW11) | is.na(ALSA_w11_dbp$DBP15W11) | 
                   is.na(ALSA_w11_dbp$DBP45W11) ) 
all_Nas_ind <- which(is.na(ALSA_w11_dbp$DBP3MW11) & 
                       is.na(ALSA_w11_dbp$DBP4MW11) & 
                       is.na(ALSA_w11_dbp$DBP5MW11) & 
                       is.na(ALSA_w11_dbp$DBP15W11) & 
                       is.na(ALSA_w11_dbp$DBP45W11) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in DBP4MW11 and DBP5MW11:
new_Nas_ind_4_5 <- new_Nas_ind[1]
ALSA_w11_dbp$mean_dbp[new_Nas_ind_4_5] <- apply(X = ALSA_w11_dbp[new_Nas_ind_4_5,c(2,5:6)], MARGIN = 1, FUN = mean)

#individuals with NA in DBP15W11 and DBP45W11:
new_Nas_ind_11_45 <- new_Nas_ind[c(2, 3, 6)]
ALSA_w11_dbp$mean_dbp[new_Nas_ind_11_45] <- apply(X = ALSA_w11_dbp[new_Nas_ind_11_45,c(2:4)], MARGIN = 1, FUN = mean)

#individuals with NA in DBP45W11:
new_Nas_ind_45 <- new_Nas_ind[4:5]
ALSA_w11_dbp$mean_dbp[new_Nas_ind_45] <- apply(X = ALSA_w11_dbp[new_Nas_ind_45,c(2:5)], MARGIN = 1, FUN = mean)



ALSA_ds_w11 <- tibble(id=ALSA_w11_dbp$id)
ALSA_ds_w11$dbp <- car::recode(ALSA_w11_dbp$mean_dbp, "NA=999")
ALSA_ds_w11$dbp <- labelled(ALSA_ds_w11$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript11}
Continuous_summary(var = ALSA_ds_w11$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w11, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation11}
```




### ALSA - Wave 12 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `DBP15W12`
* Label: `BP- Diastolic at 45 secs standing`
* Type of variable: `countinous`

* Name: `DBP45W12`
* Label: `BP- Diastolicat 105 sec standing`
* Type of variable: `countinous`

* Name: `DBP3MW11`
* Label: `BP- Diastolic at 3 mins`
* Type of variable: `countinous`

* Name: `DBP4MW12`
* Label: `BP- Diastolic at 4 mins`
* Type of variable: `countinous`

* Name: `DBP5MW11`
* Label: `BP - Diastolic at 5 mins`
* Type of variable: `countinous`


```{r assign12}
opal.assign.table.tibble(o, 'ALSA_w12_dbp','ALSA.ALSA-Wave12',
                         variables=list('DBP15W12', 'DBP45W12', 'DBP3MW12', 'DBP4MW12', 'DBP5MW12'), missings = TRUE)
```

```{r local12}
ALSA_w12_dbp <- opal.execute(o,'ALSA_w12_dbp')

Continuous_summary(var = ALSA_w12_dbp$DBP15W12, missing_values = NA)
ggplot(ALSA_w12_dbp, aes(DBP15W12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_dbp$DBP45W12, missing_values = NA)
ggplot(ALSA_w12_dbp, aes(DBP45W12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_dbp$DBP3MW12, missing_values = NA)
ggplot(ALSA_w12_dbp, aes(DBP3MW12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_dbp$DBP4MW12, missing_values = NA)
ggplot(ALSA_w12_dbp, aes(DBP4MW12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = ALSA_w12_dbp$DBP5MW12, missing_values = NA)
ggplot(ALSA_w12_dbp, aes(DBP5MW12)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the five measurements of the study-specific variable, but the missing values have to be recoded as follows:

* `NA as 999`

R code of the ALGORITHM:

```{r harmo12}
ALSA_w12_dbp$mean_dbp <- apply(X = ALSA_w12_dbp[,c(2:6)], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(ALSA_w12_dbp$DBP3MW12) | is.na(ALSA_w12_dbp$DBP4MW12) | 
                   is.na(ALSA_w12_dbp$DBP5MW12) | is.na(ALSA_w12_dbp$DBP15W12) | 
                   is.na(ALSA_w12_dbp$DBP45W12) ) 
all_Nas_ind <- which(is.na(ALSA_w12_dbp$DBP3MW12) & 
                       is.na(ALSA_w12_dbp$DBP4MW12) & 
                       is.na(ALSA_w12_dbp$DBP5MW12) & 
                       is.na(ALSA_w12_dbp$DBP15W12) & 
                       is.na(ALSA_w12_dbp$DBP45W12) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in DBP5MW12:
new_Nas_ind_5 <- new_Nas_ind[2]
ALSA_w12_dbp$mean_dbp[new_Nas_ind_5] <- apply(X = ALSA_w12_dbp[new_Nas_ind_5,c(2:3,5:6)], MARGIN = 1, FUN = mean)

#individuals with NA in DBP45W12:
new_Nas_ind_45 <- new_Nas_ind[3]
ALSA_w12_dbp$mean_dbp[new_Nas_ind_45] <- apply(X = ALSA_w12_dbp[new_Nas_ind_45,c(2:5)], MARGIN = 1, FUN = mean)

#individuals with only not NA in DBP3MW12:
new_Nas_ind_3 <- new_Nas_ind[6]
ALSA_w12_dbp$mean_dbp[new_Nas_ind_3] <- ALSA_w12_dbp$DBP3MW12[new_Nas_ind_3]


#individuals with NA in DBP15W12 and DBP45W12:
new_Nas_ind_15_45 <- new_Nas_ind[c(1,4,5)]
ALSA_w12_dbp$mean_dbp[new_Nas_ind_15_45] <- apply(X = ALSA_w12_dbp[new_Nas_ind_15_45,c(2:4)], MARGIN = 1, FUN = mean)

ALSA_ds_w12 <- tibble(id=ALSA_w12_dbp$id)
ALSA_ds_w12$dbp <- car::recode(ALSA_w12_dbp$mean_dbp, "NA=999")
ALSA_ds_w12$dbp <- labelled(ALSA_ds_w12$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript12}
Continuous_summary(var = ALSA_ds_w12$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(ALSA_ds_w12, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation12}
```




### ALSA - Wave 13 

There is no variable.


# Quality estimation
Calculate the mean of the 5 variables used in all waves of the study.


```{r save, echo=FALSE}
dbp <- ALSA_ds_w7
save(dbp, file = "../RData/w7/dbp.RData")
rm(dbp)

dbp <- ALSA_ds_w9
save(dbp, file = "../RData/w9/dbp.RData")
rm(dbp)

dbp <- ALSA_ds_w11
save(dbp, file = "../RData/w11/dbp.RData")
rm(dbp)

dbp <- ALSA_ds_w12
save(dbp, file = "../RData/w12/dbp.RData")
rm(dbp)
```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
