==> ALSA/6_Physical_measures/12_Arterial_hypertension.Rmd <==
---
title: "Arterial hypertension"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `ah`
* Variable label: `Arterial hypertension`
* Variable description: `Arterial hypertension. Means of all diastolic blood pressure measures.`
* Domain: `Physical Measures`
* Value type: `Categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | no hypertension
1        | hypertension

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Arterial hypertension"
ds_label_all <- "Arterial hypertension. Means of all diastolic blood pressure measures."
```

# Data process

## ALSA

### Wave 1 


There is no variable.




### Wave 2 

There is no variable.ol




### Wave 3 

There is no variable. 




### Wave 4 

There is no variable.





### Wave 5 

There is no variable.






### Wave 6 

There is no variable.






### Wave 7 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | 
| | `992 = Impute` | `992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable`| `996 = Not applicable`| 
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` | `998 = Do not know` |
| | `999 = Missing`| `999 = Missing`|
| **Description** |  |  |


```{r assign7, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w7/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w7/dbp.RData")
ALSA_ah <- merge(sbp, dbp, by = "id")
```

```{r local7, echo=FALSE}
vari <- ALSA_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- ALSA_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo7}
ALSA_ah$dh <- car::recode(ALSA_ah$dbp, "0:89.99999999=0; 90:900=1")
ALSA_ah$sh <- car::recode(ALSA_ah$sbp, "0:139.9999999=0; 140:900=1")

ALSA_ds_w7 <- tibble(id=ALSA_ah$id)
ALSA_ds_w7$ah <- NA
ALSA_ds_w7$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==0)] <- 0
ALSA_ds_w7$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==0)] <- 1
ALSA_ds_w7$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w7$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w7$ah[which(ALSA_ah$dh==999 & ALSA_ah$sh==999)] <- 999

ALSA_ds_w7$ah <- labelled(ALSA_ds_w7$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript7, echo=FALSE}
vari <- ALSA_ds_w7$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(ALSA_ds_w7, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation7, echo=FALSE}
```

```{r importOPAL7, echo=FALSE}
#val_labels(ALSA_ds_w7) <- NULL
#opal.assign.data(o, 'ALSA_ds_w7', ALSA_ds_w7)
#opal.symbol_import(o,'ALSA_ds_w7', project='_Athlos_Harm_Dataset')
ah <- ALSA_ds_w7
save(ah, file = "../RData/w7/ah.RData")
rm(ah)
```


### Wave 8 

There is no variable.






### Wave 9 

#### Study-specific variable description


| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | 
| | `992 = Impute` | `992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable`| `996 = Not applicable`| 
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` | `998 = Do not know` |
| | `999 = Missing`| `999 = Missing`|
| **Description** |  |  |

    
```{r assign9, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w9/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w9/dbp.RData")
ALSA_ah <- merge(sbp, dbp, by = "id")
```

```{r local9, echo=FALSE}
vari <- ALSA_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- ALSA_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo9}
ALSA_ah$dh <- car::recode(ALSA_ah$dbp, "0:89.99999999=0; 90:900=1")
ALSA_ah$sh <- car::recode(ALSA_ah$sbp, "0:139.9999999=0; 140:900=1")

ALSA_ds_w9 <- tibble(id=ALSA_ah$id)
ALSA_ds_w9$ah <- NA
ALSA_ds_w9$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==0)] <- 0
ALSA_ds_w9$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==0)] <- 1
ALSA_ds_w9$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w9$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w9$ah[which(ALSA_ah$dh==999 & ALSA_ah$sh==999)] <- 999
ALSA_ds_w9$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==999)] <- 999

ALSA_ds_w9$ah <- labelled(ALSA_ds_w9$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript9, echo=FALSE}
vari <- ALSA_ds_w9$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(ALSA_ds_w9, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation9, echo=FALSE}
```

```{r importOPAL8, echo=FALSE}
ah <- ALSA_ds_w9
save(ah, file = "../RData/w9/ah.RData")
rm(ah)
```



### Wave 10 

There is no variable.





### Wave 11 

#### Study-specific variable description


| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | 
| | `992 = Impute` | `992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable`| `996 = Not applicable`| 
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` | `998 = Do not know` |
| | `999 = Missing`| `999 = Missing`|
| **Description** |  |  |

    
```{r assign11, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w11/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w11/dbp.RData")
ALSA_ah <- merge(sbp, dbp, by = "id")
```

```{r local11, echo=FALSE}
vari <- ALSA_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- ALSA_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo11}
ALSA_ah$dh <- car::recode(ALSA_ah$dbp, "0:89.99999999=0; 90:900=1")
ALSA_ah$sh <- car::recode(ALSA_ah$sbp, "0:139.9999999=0; 140:900=1")

ALSA_ds_w11 <- tibble(id=ALSA_ah$id)
ALSA_ds_w11$ah <- NA
ALSA_ds_w11$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==0)] <- 0
ALSA_ds_w11$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==0)] <- 1
ALSA_ds_w11$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w11$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w11$ah[which(ALSA_ah$dh==999 & ALSA_ah$sh==999)] <- 999

ALSA_ds_w11$ah <- labelled(ALSA_ds_w11$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript11, echo=FALSE}
vari <- ALSA_ds_w11$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(ALSA_ds_w11, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation11, echo=FALSE}
```

```{r importOPAL10, echo=FALSE}
ah <- ALSA_ds_w11
save(ah, file = "../RData/w11/ah.RData")
rm(ah)
```


### Wave 12 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | 
| | `992 = Impute` | `992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable`| `996 = Not applicable`| 
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` | `998 = Do not know` |
| | `999 = Missing`| `999 = Missing`|
| **Description** |  |  |

    
```{r assign12, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w12/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/ALSA/RData/w12/dbp.RData")
ALSA_ah <- merge(sbp, dbp, by = "id")
```

```{r local12, echo=FALSE}
vari <- ALSA_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- ALSA_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(ALSA_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo12}
ALSA_ah$dh <- car::recode(ALSA_ah$dbp, "0:89.99999999=0; 90:900=1")
ALSA_ah$sh <- car::recode(ALSA_ah$sbp, "0:139.9999999=0; 140:900=1")

ALSA_ds_w12 <- tibble(id=ALSA_ah$id)
ALSA_ds_w12$ah <- NA
ALSA_ds_w12$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==0)] <- 0
ALSA_ds_w12$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==0)] <- 1
ALSA_ds_w12$ah[which(ALSA_ah$dh==0 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w12$ah[which(ALSA_ah$dh==1 & ALSA_ah$sh==1)] <- 1
ALSA_ds_w12$ah[which(ALSA_ah$dh==999 & ALSA_ah$sh==999)] <- 999

ALSA_ds_w12$ah <- labelled(ALSA_ds_w12$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript12, echo=FALSE}
vari <- ALSA_ds_w12$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(ALSA_ds_w12, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation12, echo=FALSE}
```

```{r importOPAL12, echo=FALSE}
ah <- ALSA_ds_w12
save(ah, file = "../RData/w12/ah.RData")
rm(ah)
```


### Wave 13 

There is no variable.

# Quality estimation

No comments.





```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
